pyproject.toml
[project]
name = "TF-Restormer"
version = "0.1.0"
description = "PyTorch implementation of TF-Restormer"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "beartype>=0.21.0",
    "colorednoise==2.2.0",
    "einops==0.8.1",
    "loguru==0.7.3",
    "matplotlib==3.10.3",
    "pedalboard==0.9.17",
    "ptflops==0.7.4",
    "seaborn==0.13.2",
    "tensorboard>=2.4.0",
    "tensorboardx==2.6.4",
    "thop==0.1.1.post2209072238",
    "pesq>=0.0.4",
    "torch-pesq==0.1.2",
    "torch-stoi>=0.2.3",
    "torchinfo==1.8.0",
    "transformers==4.53.3",
    "utmosv2",
    "wandb==0.21.0",
    "soundfile==0.13.1",
    "librosa==0.11.0",
    "scipy==1.16.0",
    "lightning>=2.5.2",
    "ipykernel>=6.30.1",
    "notebook>=7.4.5",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
    "pytorch-lightning>=2.5.2",
    "fastdtw>=0.3.4",
    "onnxruntime>=1.20.1",
    "python-dotenv>=0.19.0",
    "rotary-embedding-torch==0.8.8",
    "tqdm==4.67.1",
    "pandas==2.3.1",
    "pyyaml==6.0.2",
    "scikit-learn==1.7.1",
    "annotated-types==0.7.0",
    "audioread==3.0.1",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "filelock==3.18.0",
    "fonttools==4.59.0",
    "fsspec==2025.7.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "hf-xet==1.1.5",
    "huggingface-hub==0.33.4",
    "idna==3.10",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "kiwisolver==1.4.8",
    "lazy-loader==0.4",
    "llvmlite==0.44.0",
    "markupsafe>=2.1.0",
    "mpmath==1.3.0",
    "msgpack==1.1.1",
    "networkx==3.5",
    "numba==0.61.2",
    "packaging==25.0",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "pooch==1.8.2",
    "protobuf==6.31.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "pystoi==0.4.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "safetensors==0.5.3",
    "sentry-sdk==2.33.2",
    "setuptools==78.1.1",
    "six==1.17.0",
    "smmap==5.0.2",
    "soxr==0.5.0.post1",
    "sympy>=1.13.3",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.2",
    "torchtyping==0.1.5",
    "triton==3.3.1",
    "typeguard==2.13.3",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wheel==0.45.1",
    "nltk>=3.9.1",
    "levenshtein>=0.27.1",
    "jellyfish>=1.2.0",
    "numpy>=2.2.6",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "pysptk>=1.0.1",
    "pyworld>=0.3.5",
    "ninja>=1.13.0",
    "causal-conv1d>=1.5.2",
    "mamba-ssm>=2.2.5",
    "mir-eval>=0.8.2",
]

[tool.uv]
index-strategy = "unsafe-best-match"

[tool.uv.sources]
utmosv2 = { git = "https://github.com/sarulab-speech/UTMOSv2.git" }

[tool.uv.pip]
extra-index-url = ["https://download.pytorch.org/whl/cu124"]
